Tenant and Automat services, controllers, repositories - need to fix tenantId

+++ magazynV3/src/main/java/io/agileintelligence/ppmtool/repositories/TenantRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface TenantRepository extends CrudRepository<Tenant, Long> {
-    Tenant findByNip(Long nip);
+    Tenant findByNip(Integer nip);
 
 
 }
